+++ lib/features/braews_view.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class BrawesView extends StatelessWidget {
+  const BrawesView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
