+++ lib/features/splash/pages/splash_view.dart
