+++ lib/features/wishlist_view.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class WishListView extends StatelessWidget {
+  const WishListView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
